pyproject.toml
[project]
name = "sndtk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pathspec>=0.12.1",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.5",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501", # line-too-long (handled by formatter)
    "B006",
]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
